FINISHED
Updated the scroll pane with a list instead of buttons. Enabled the
ability to add contacts, remove contacts and edit contacts with the
changes appearing in the list. Enabled the ability to sort and search
with the changes appearing in the list. Got rid of the file system
class and the referencing button class.